chore(meta): remove unused DFS test cases; mv snapshot test to standalone file
