+++ src/meta/raft-store/src/state_machine/testing.rs
@@ -118,52 +118,3 @@ pub fn cases_incr_seq() -> Vec<(&'static str, Option<RaftTxId>, &'static str, u6
         ("no txid, no de-dup", None, "k2", 2),
     ]
 }
-
-pub type AddFileCase = (
-    &'static str,
-    Option<RaftTxId>,
-    &'static str,
-    &'static str,
-    Option<String>,
-    Option<String>,
-);
-
-// test cases for Cmd::AddFile
-// case_name, txid, key, value, want_prev, want_result
-pub fn cases_add_file() -> Vec<AddFileCase> {
-    vec![
-        (
-            "add on none",
-            Some(RaftTxId::new("foo", 1)),
-            "k1",
-            "v1",
-            None,
-            Some("v1".to_string()),
-        ),
-        (
-            "add on existent",
-            Some(RaftTxId::new("foo", 2)),
-            "k1",
-            "v2",
-            Some("v1".to_string()),
-            None,
-        ),
-        (
-            "dup set with same serial, even with diff key, got the previous result",
-            Some(RaftTxId::new("foo", 2)),
-            "k2",
-            "v3",
-            Some("v1".to_string()),
-            None,
-        ),
-        (
-            "diff client, same serial",
-            Some(RaftTxId::new("bar", 2)),
-            "k2",
-            "v3",
-            None,
-            Some("v3".to_string()),
-        ),
-        ("no txid", None, "k3", "v4", None, Some("v4".to_string())),
-    ]
-}
